Separate v2 validity and task constraint checkers
