+++ src/retrocast/v2/workflow/__init__.py
@@ -10,9 +10,7 @@ from retrocast.v2.workflow.collect import (
 from retrocast.v2.workflow.ingest import ingest_candidates, ingest_routes
 from retrocast.v2.workflow.score import (
     ConstraintChecker,
-    TaskConstraintChecker,
     TierChecker,
-    TierZeroChecker,
     score,
     score_candidate,
     score_target,
@@ -22,9 +20,7 @@ __all__ = [
     "CollectedCandidates",
     "CollectedRoutes",
     "ConstraintChecker",
-    "TaskConstraintChecker",
     "TierChecker",
-    "TierZeroChecker",
     "adapt_candidates",
     "adapt_route",
     "adapt_routes",
