+++ tests/v2/workflow/test_score.py
@@ -5,16 +5,49 @@ from retrocast.typing import ErrorCode, InChIKeyStr, SmilesStr
 from retrocast.v2.models import (
     Candidate,
     CheckStatus,
+    ConstraintResult,
     FailureRecord,
     Molecule,
     Reaction,
+    ReactionValidity,
     Route,
+    RouteValidity,
     Target,
     Task,
     TaskConstraints,
     Tier,
+    TierResult,
 )
-from retrocast.v2.workflow.score import TaskConstraintChecker, TierZeroChecker, score, score_candidate
+from retrocast.v2.workflow.score import score, score_candidate
+
+
+class FixedTierChecker:
+    tier = Tier.ZERO
+    name = "fixed-tier-zero"
+
+    def __init__(self, status: CheckStatus = CheckStatus.PASS) -> None:
+        self.status = status
+
+    def check_route(self, route: Route) -> RouteValidity:
+        reaction_tiers = {}
+        try:
+            reaction_id = route.reaction_at("rc:r:/").id()
+        except KeyError:
+            reactions = []
+        else:
+            reaction_tiers[Tier.ZERO] = TierResult(status=self.status)
+            reactions = [ReactionValidity(reaction_id=reaction_id, tiers=reaction_tiers)]
+        return RouteValidity(tiers={Tier.ZERO: TierResult(status=self.status)}, reactions=reactions)
+
+
+class FixedConstraintChecker:
+    name = "fixed-task"
+
+    def __init__(self, status: CheckStatus = CheckStatus.PASS) -> None:
+        self.status = status
+
+    def check_route(self, route: Route, constraints: TaskConstraints) -> ConstraintResult:
+        return ConstraintResult(status=self.status)
 
 
 def molecule(smiles: str, *, product_of: Reaction | None = None) -> Molecule:
@@ -33,24 +66,6 @@ def route(target_smiles: str = "CCO", reactant_smiles: tuple[str, ...] = ("C", "
             product_of=Reaction(reactants=[molecule(smiles) for smiles in reactant_smiles]),
         )
     )
-
-
-def stock_for(test_route: Route) -> set[InChIKeyStr]:
-    return {InChIKeyStr(get_inchi_key(smiles)) for smiles in leaf_smiles(test_route)}
-
-
-def leaf_smiles(test_route: Route) -> list[str]:
-    leaves = []
-
-    def visit(molecule: Molecule) -> None:
-        if molecule.product_of is None:
-            leaves.append(molecule.smiles)
-            return
-        for reactant in molecule.product_of.reactants:
-            visit(reactant)
-
-    visit(test_route.target)
-    return leaves
 
 
 def target(acceptable_routes: list[Route] | None = None) -> Target:
@@ -77,8 +92,8 @@ def test_failed_candidate_does_not_satisfy_solv_zero() -> None:
         candidate,
         target=target(),
         constraints=TaskConstraints(),
-        tier_checkers=[TierZeroChecker()],
-        constraint_checker=TaskConstraintChecker(),
+        tier_checkers=[FixedTierChecker()],
+        constraint_checker=FixedConstraintChecker(),
     )
 
     assert scored.failed_adaptation()
@@ -88,13 +103,12 @@ def test_failed_candidate_does_not_satisfy_solv_zero() -> None:
 
 
 def test_valid_route_can_satisfy_tier_zero_and_task() -> None:
-    test_route = route()
     scored = score_candidate(
-        Candidate(rank=2, route=test_route),
+        Candidate(rank=2, route=route()),
         target=target(),
-        constraints=TaskConstraints(stock="stock-a"),
-        tier_checkers=[TierZeroChecker()],
-        constraint_checker=TaskConstraintChecker(stock=stock_for(test_route), stock_name="stock-a"),
+        constraints=TaskConstraints(),
+        tier_checkers=[FixedTierChecker()],
+        constraint_checker=FixedConstraintChecker(),
     )
 
     assert scored.has_route()
@@ -108,8 +122,8 @@ def test_task_constraint_failure_prevents_solv_zero() -> None:
         Candidate(rank=1, route=route()),
         target=target(),
         constraints=TaskConstraints(stock="stock-a"),
-        tier_checkers=[TierZeroChecker()],
-        constraint_checker=TaskConstraintChecker(stock=set(), stock_name="stock-a"),
+        tier_checkers=[FixedTierChecker()],
+        constraint_checker=FixedConstraintChecker(CheckStatus.FAIL),
     )
 
     assert scored.satisfies_validity(Tier.ZERO)
@@ -122,45 +136,13 @@ def test_reaction_validity_is_addressable_by_reaction_id() -> None:
         Candidate(rank=1, route=route()),
         target=target(),
         constraints=TaskConstraints(),
-        tier_checkers=[TierZeroChecker()],
-        constraint_checker=TaskConstraintChecker(),
+        tier_checkers=[FixedTierChecker()],
+        constraint_checker=FixedConstraintChecker(),
     )
 
     result = scored.reaction_tier_result("rc:r:/", Tier.ZERO)
     assert result is not None
     assert result.status == CheckStatus.PASS
-
-
-def test_empty_reaction_fails_tier_zero() -> None:
-    invalid_route = Route(target=molecule("CCO", product_of=Reaction(reactants=[])))
-
-    scored = score_candidate(
-        Candidate(rank=1, route=invalid_route),
-        target=target(),
-        constraints=TaskConstraints(),
-        tier_checkers=[TierZeroChecker()],
-        constraint_checker=TaskConstraintChecker(),
-    )
-
-    assert not scored.satisfies_validity(Tier.ZERO)
-    result = scored.reaction_tier_result("rc:r:/", Tier.ZERO)
-    assert result is not None
-    assert result.status == CheckStatus.FAIL
-
-
-def test_inchikey_mismatch_fails_tier_zero() -> None:
-    invalid_route = Route(target=Molecule(smiles=SmilesStr("CCO"), inchikey=InChIKeyStr(get_inchi_key("C"))))
-
-    scored = score_candidate(
-        Candidate(rank=1, route=invalid_route),
-        target=target(),
-        constraints=TaskConstraints(),
-        tier_checkers=[TierZeroChecker()],
-        constraint_checker=TaskConstraintChecker(),
-    )
-
-    assert not scored.satisfies_validity(Tier.ZERO)
-    assert scored.tier_result(Tier.ZERO).checks[0].code == "tier0.inchikey_mismatch"
 
 
 def test_score_preserves_candidate_rank_and_records_acceptable_match() -> None:
@@ -170,8 +152,8 @@ def test_score_preserves_candidate_rank_and_records_acceptable_match() -> None:
     evaluation = score(
         {"ethanol": [Candidate(rank=7, route=predicted_route)]},
         task(benchmark_target),
-        tier_checkers=[TierZeroChecker()],
-        constraint_checker=TaskConstraintChecker(),
+        tier_checkers=[FixedTierChecker()],
+        constraint_checker=FixedConstraintChecker(),
     )
 
     scored = evaluation.targets["ethanol"].candidates[0]
